Update 'playGame' function and how it is called
The function was updated so that it takes the DOM nodes it uses as
parameters. This function is used in the r-p-s buttons event listeners.
For it to work properly an anonymous function was passed to the event
listener and inside that one, playGame was called, which allowed me to
call it with the corresponding parameters.